requirements.txt
numpy==1.26.1
pandas==2.1.2
scikit-learn==1.3.2
pyarrow==14.0.2
Flask==3.0.1
matplotlib==3.8.2
mysql-connector-python==8.3.0